requirements.txt
tkinter == 8.6.13
paho-mqtt == 2.1.0
numpy  == 1.26.0